I've implemented a system global Unix time stamp -- g_unixTime. This is a second counter register, holding the number of seconds since 1970-01-01 00:00:00. To save program code space, I've done away with time zone and daylight savings processing, which seemed perfectly reasonable for our purposes.
The code takes up about 5Kbytes of Flash ROM, which is not trivial. But I think
the flexibility it gives for the future is worth it. With this, it will now be
easy to do date and time based calculations or whatever. More importantly, this
method consumes the least possible CPU time to do each second's increment,
since it doens't have to worry about crossing midnight, how many days in
the current month or whether its a leap year -- not during the increment itself
anyway. Those considerations can happen only as needed, in low priorty code 
space.

For want of a better place to put it, I've added a 4th "DATE AND TIME" screen
to the FrSky menus. At the moment, this merely shows the current date and time.
But the intention is to make this screen the place to set the date/time, also.

So right now, the real time clock chip (on the SD card interface) is being read
once, at power-up. The individual year/month/day/hour/min/sec data from that is
computed down to a Unix time stamp, as described above and used to initialise
g_unixTime.  From then on, g_unixTime is incremented by one, once every 100ms,
from inside the existing per10ms() function.

The above also means that any time an SD card file date/time needs updating,
the system will not have to go through the CPU and time intensive process
of extracting the data from the RTC chip. This is good!

Of course, none of this appears in the stock hardware version of the code and
only works with the prototype V3 PC board. But it's all good foundational work
for moving on to more interesting and exciting things, soon!