Fixed keyDown() to be able to handle the V3 key matrix circuitry.
We want to get rid of the CPU loading "scanned key matrix" thing for the V4
PCB, since there will again be plenty of spare I/O ports to allow it. Cam is
working on it. Might be a challenge -- but it's important if the board is to
become a truly "drop in replacement". Otherwise, you have to hack the trim
switch boards and add new wires to make the board work. (This is still the case
with the V3 board -- leaving me to wonder if it's even worth releasing.
Probably not. :-/)