Bandgap issue solved.
Copy of reply to my own post at AVRfreaks.com ...

OK, I've found the solution to this problem, though it's not as pretty as I'd
hoped.

The bottom line appears to be that the ATmega2561 chip's 1.1V internal
reference source has a much higher impedance than the 1.22V reference circuitry
in the ATmega64A. In the end, inserting a delay of no less than 300us
(micro-seconds) after the MUX selection gave us consistently stable Vbg bandgap
samples.

Conversely, the ATmega64A will return stable samples with only 5us delay -- or
simply taking two samples and discarding the first, as Atmel suggests.

Interestingly, on two separate units, one with a 64A and one with a 2561, the
stable Vbg samples are both exactly 241. This has to be a mere fluke, since in
fact the 2561 chip would ideally return 225 and the 64A 250 -- assuming a
perfect 5.0V AVCC reference and perfect 1.1V and 1.22V bandgap references,
respectively. *shrug*