Experimental easy change of cAsE when entering model names.
The idea is that, instead of having to auto-repeat a long while to change say,
a G to a g, you can now put the cursor over the G and do a [RIGHT/LEFT LONG]
(what used to be auto-repeat left and right, along the name, which for such
short words, I figured should not be missed.)

A side-effect is that the cursor does not move left or right until the
[LEFT/RIGHT] keys are released. But it doesn't seem so bad.

Bertrand, if you like this, perhaps you can add it to flight phase name entry
too? If you don't like it, please say so. But it would be nice to have some way
of changing letter cAsE without having to scroll so far and awkwardly. I could
not think of any other key_press event to make it happen by.

Cam -- your comments also welcomed, of course!

This was also the first coding I did where I tested it using only simu. Very handy!