Some optimisations to my (Gruvin) recent Fr-Sky user data experimental code. Still seeing jitter.
I've double-checked with an old transmitter (v2.14 hacked PCB) that jitter is
not a problem on that version, which contains Bertrand's per-model fr-sky
setting etc. ON that device, jitter is 0 -- min/max latency = 6 -- with
telemtry data flowing.

I've done some minor optimisation of some of Bertrand's frsky related code to
reduce some cpu load and stack usage. It made no difference.

So the mystery as to what is causing jitter now continues. :/