Jitter problem solved. It was really simply in the end and I feel like a twit.
I didn't realise that the MCU automatically set the global interrupt flag when
it vectored to a service routine. All I had to do was issue an sei to re-enable
interrupts as soon as the USART data register had been cleared.

So we're back to nearly zero jitter and still seem to have nice smooth
stick-to-servo control -- even while fr-sky user data is streaming in at full
speed. This is much more as I originally expected and I'm certainly breathing
sighs of relief!