More SD/MMC card logging work.
Implemented automatic log filenames and creation of /G9XLOGS folder, should it
not exist.

Cam helped me decide how all this will work (tahnks again Cam!) ...

The log filename is decided when a new model is loaded -- either at power on or
if you select a new model yourself.

Log files are not actually created until the LOG switch is activated.

Logs are stored in /G9XLOGS and have a filename format ...

  Mxx_nnn

... where xx is the currently loaded model number (00 to 16) and nnn is a
number that increments from 000 to a potential 999 (Heavin forbid 999 is ever
reached. But if so, that file will be overwritten after power cycle or new
model selection.)

Files have valid creation date/time stamps to aid in the obvious manne.

Note that 'nnn' only increments when a new model is chosen or if the power is
cycled -- and only if a file of the preceeding 'nnn' number already existed.
Thus each activation of the LOG switch will continue to append log data to the
same model file. This means you can turn logging off and back on during a
flight and all the data will end up in the same file, essentially for that
flying session.

The "LOG switch" is currently hard coded to be one and the same as the LCD
back-light. Clearly, this needs addressing with its own menu configurion
option, so that we can define any switch (including custom types) to turn
logging on or off.

The logged data itself is still only raw Fr-Sky Data. I have ordered a Fr-Sky
telemetry hub with some sensors and intend logging CSV-file text from data
collected via Fr-Sky's proprietary (but published) hub protocol. Something like
...

0,90,-174.2342,-35.5346,10:11:22,1,0,1,0,1, etc

The fields will have a heading row with their respective names, which will
include the status of the 8 9X physical switches and perhaps even custom
switches and the like. (We don't want to big the CPU down with too much data!
This is something I'll have to test and play by ear as we go.)

The first field in the CSV file will be a number that increments each time the
LOG switch is activated. This is intended to make separating single-flight log
segments easy, without having to study date/time stamps or whatever.