More telemetry display code completed.
Adjusted font_dblsize.xmb down one pixel, so that the bottom of double-size
characters align with the bottom of normal size characters when displayed
alongside eachother. This also fixes the eye-sore missing top row when
double-size characters are printed INVERS.

Fixed lcd_lastPos (last X position drawn to by fount output etc) so it worked
properly throughout, ... which led to getting utterly frustrated with somewhat
messy and annoyingly buggy lcd_outdezNAtt functioni -- for the LAST time. 

So, I re-wrote the whole thing from scratch. The resulting code is a fair bit
smaller and more efficient, at the cost of a small (non-static) 5 byte character 
buffer.

When writing the output for the telemtry GPS values, I needed to be able to pad
more than one 0 both to the right and to the left of a printed number. So
*PLEASE NOTE* that I've removed 'LEADING0' from the mode flags register and set
it up, along with the new TRAILING0 flag to be logcially ORed (or added) to
'len', instead. 

Both LEADING0 and TRAILING0 now pad all the way out to 'len' characters. For
this work, len now has a functional maximum input value of 0x3f (63), which is
much wider than the LCD display in characters units. 

I've also re-instated the SIGN and PREC3 flags (while keeping within 8-bits total)
This was to allow 3 decimal places (yes -- I want that!) and to allow unsigned
integers up to 65535 to be handled.

Finally, the default RIGHT alignment alligns ON THE DECIMAL POINT -- as it was
originally designed -- NOT the right-hand edge of the last displayed character. 

Here are some examples ...

Example 1:

  lcd_outdezNAtt(x, y, 1234, PREC2, 5|TRAILING0);

  --> 12.340

Example 2:

  lcd_outdezNAtt(x, y, 32, PREC2, 4|LEADING0);

  --> 0032

Example 3:

  lcd_outdezNAtt(x+20, y, 46573, PREC3, 5|TRAILING0);
  lcd_outdezNAtt(x+20, y+FH, 1234, PREC2, 5|TRAILING0);

  -->   46.573
  -->    1.2300
  Note that those two numbers are aligned on the decimal point. Also note that
  TRAILING0 pads zeros using digit-counts, not chacater count. That is, the
  decimal point itself does not count as a digit count space (nor does it take
  up a character space on the actual 9X implemention.)

Example 4:

  lcd_outdezNAtt(x+20, y, 46573, PREC3|LEFT|SIGN, 5|TRAILING0);
  lcd_outdezNAtt(x+20, y+FH, 1234, PREC2|LEFT, 5|TRAILING0);

  -->    -46.573
  -->    1.2300