Second page of Fr-Sky telemetry display completed, bringing the display phase to and (almost) end.
For some reason, the barometric pressure sensor (altitude) value is being
reported as 65390, when the range is specified as 0-9999. Surely it can only be
coincidence that the inverted and reversed BCD of that number is 19m, which
would be about correct. :P Hoenstly, The sensor does respond to artifical
pressure increases and the number seems to move by about the right amount, in
meters. But really, I have no idea what the problem is.