PCBV4 ADC stuff all sorted.
So all "standard" functions are now working on the V4 board. By "standard", I
mean switches and ADC inputs.

Oh and, luckily enough, PPM_out is working and servo move as they should too!

   **** A BIG   C O N G R A T U L A T I O N S  to Cameron for a job well done! ****

Of course, there's a fair more testing to be done yet before we're sure that
ALL of the board is working right. But at least we know all the main stuff is
good. Whew!

- - - -

We will now need hack all the Beta v4.0 boards to make them into v4.1 boards,
   electrically. This is needed bcause we don't want to have to mess about
   having a v4.0 and a v4.1 in the firmware.

As far as I'm aware, the only hack needed at this stage is for the hardware
PPM.

HOWEVER -- let's hold of on any hacking until at least Cam and I have had a
chance to get our v4 boards built and tested. Hmmm... I believe Brano is also
in the midst of construction. Brano -- if you're reading this, please email me
or the groupd wth a progress report.