Added the two custom library items for sdcard.sch.
CAM: How did you get all you components in a single library
(share/libraries/gruvin.lib) -- and how are you accessing them, when that
library is not even loaded? I cannot find any way in Library Editor to open a
library file that is not already included as part of the project, in the
Preferences area.