Figured out how to get my custom components incorporated into Cam's existing gruvin.lib. (KiCAD)
It's a bit vague though, aint it? Doesn't seem to be any easy way to copy
components to and from other libraries. You seem to have to create a new
component, based on the one in a separate library, then save it to the
"current" library -- and do that for each individual component you want to
mearge into the one common library. *shrug* Still, it beats having one file 
per component, like I had before!