Work on getting the external 1.2V voltage reference to work in place of the internal bandgap reference.
It's all working. But the math is horrendously horrible.

+ I wanted 2 decimal plces of accuracy for the batt volts calibration.  + I
wanted to be able to adjust the calibration in as near to 0.01V steps as
possible (practical) + I wanted the single decimal place main display of
voltage to round up and down "properly"

I have achieved all of that. But it aint pretty.

The problem is that with a 1.2V reference, it just so happens that there's no
small numbers fraction that computes to an accurate enough value, IMO. So I has
to increase the likes of 18/256 to more like 139/2048. Then that had to
multiplied by ten, to get my extra decimal place. Then the math to do the
rounding, neccesarily starting at these big values involved an even bigger
value. (At least it all fits inside 32 bit easily, without have to go 64-bit!!
             :{ Good grief.)