Tidied up Batt Volts display and calibration -- V4 board only. (Didn't want complaints for existing V1/2/3 board users.)
Calibration is now accurate to 2 decimal places (I believe this is needed for
certain battery types that are in use.) Also, the main display is properly 
rounded to 1 decimal place, instead of being truncated, as it was previously.

Yes, this requires 32-bit math to function. We can live with it. It's not the
end of the world. Thank you.

You may notice that I've kind of over-killed the reading of the external 1.2V
reference. The math for the above required the old "BandGap" be multiplied by
8. So instead of just doing that, I accumulate 4 x 10-bit samples, then
multiple by 2 to make the fnal result like the 11-bit, over-sampled value used
everywhere else. I could have sampled it 8 times, accumulatively. But that just
seemed silly, especially considering ONE sample was already enough to get a rock
solid reading! Oh well.  Better safe than sorry -- and I am sick to death of the
battery volts issue cropping up again and again, after we think we've fix it i
already. So this really should get it, once and for all.