Rotary encoder code coded. (Who invented English anyway?)
However, none of the interrupt-on-edge changes are actually firing and I have
no idea why.

I've checked the hardware. It's fine. (Cam -- the wiring you did on your
encoders, as shown in the photos you sent me, is either wrong or my dyslexia sturck
again. On my ones, which I thought I copied from yours, the push button and
one of the encoder outputs is swapped -- pins 2 and 4 swapped.)

I've checked that the encoder signals are in fact making it to the exact right
pins on the MCU. They are. (My scope probes are very fine and sharp!)

For some reason, the external interrupts just never fire when the pin inputs
change. Can anyone see WTF I've done wrong? It looks @#$! perfect to me.