Very first draft of the new Rx de-multiplexer board. This is designed for model boat people to enable them to split one channel into eight at the expense of latency.
It has 8 outputs to drive servos, and 8 500mA digital outputs (of which 4 are PPM capable). It will only be an 8 channel de-multiplexer so servo output 1 and digital output 1 will be controlled by the same input, etc.

I have done it this way to avoid having to have a de-multiplexer board for digital and one for servos.

You can drive the digital and servo outputs from the Rx battery or another battery, selected via switches.

This design may change a lot at this stage, it needs to be tested, and so far no firmware exists !

TODO: Add vias and fuses I think, also a 328 is probably over the top.