Spent ages re-selecting nz.element14.com (Farnell) brands and part codes to achieve best possible pricing for in-stock (next day delivery) parts to New Zealand.
In the end, the pricing still works out at quite a lot more then DigiKey.com for a
single board build. (Note that this is mostly due to the MOQ's on the JST
connectors). 

Total component pricing when ordering for ten boards  (MOQ's then mostly out of
the picture) works out around NZ$93 for each board.

Oh and shipping is free for online orders.