Folder renamed from archive/pcb-v4.1-frozen to archive/pcb-v4.1a-frozen
This from the new README file ...

This folder contains the original pcb-v4.1-frozen files, plus a nuber of miniro
changes that were made after the 'freeze' date to correct problems pointed out
by the assembly factory. Nothing in the actual circuit schematic or track/board
layout was changed -- only minor component properties, to do with robotic
placement cooordinates and the like.

The folder's name was changed to 4.1a as a safe-guard, just in case others has
downloaded the original frozen folder and did not realise changes had been
made.