MAIN PCB v4.2 -- Changes to PORTH header.
I have successfully changed the PORTH header from 8-pin to 10-pin and connected
GND/VCC to the extra pins. I also rotated the module so that pin-1 is now on
GND, rather than PH7. 

GND is on pad 1
VCC is on pad 2
PH0 to PH7 are on pads 3 through 10, respectively.