MAIN BOARD v4.2
A few minor tweaks -- some ground plane vias fixed. No doubt still some minor
issues to be found and rememedied.

Right now, I'm now at all confident to run this new version straight to the
factory.  With the number of changes, the potential for an unseen error is just
too great. So I need to push for a v4.2-frozen over the coming week and get
some boards ordered, for a manual build.

Yay \o/ :/ At least I'm better set up for it now, compared to in the past. Will
still be agood 5 to 7 hours though. So tempting to just trust my work and go to
production.

Any votes the other way, anyone? We've actaully never had a board design fault
in the past -- only schematic changes, which were of course reflected in the
board. On that basis, I should just trust KiCAD and our history and go for it.
But $2,000 is a lot to screw around with. Even then, I guess the worst case
scenario would be having to manually wire-patch 30 or so boards. Hmmm.

Bryan.