Fixed watch dog timer reset event's failure to reboot properly. (This remains a problem in the current 'stable' release and probably open9x as well.)
The cause of the WDT reset fault turned out to be how the WDT is automatically
reset to "active with minimum time period" when the reset occurs. This resulted
in the WDT reset re-firing continuously, before we got around to setting it to
500ms period. I guess the last time I tested this, there must have been not
quite so many C variables begin cleared and what have you, so we snuck in under
the time limit. (It was a long time ago.)

Started soft-off implementation. It just powers off immediately at present. I
haven't looked at how it was done in open9x. Figured I'd just do it my way.
*shrug*