Soft-off fully implemented. Features include ...
+ Power switch indicator flags are debounced.

+ "MESSAGE -- Powering off ..." displays for 1.3s.

+ User can abort the power down by reconnecting/switching power 
  (full WDT timer reset used to ensure full system initialisation). 

+ Watchdog timer allows 2 seconds before giving up waiting for file
  closure, to ensure power goes off no matter.

CAVEAT: EEPROM writes are not actually checked for completion. Rather, it is
assumed that 1.3 seconds will always be long enough. I'd like to improve this
-- because I know that copying large models for example can take longer -- but
don't yet know ow to check for EEPROM write activity.

BTW -- I didn't copy the open9x implementation of this only because it would have
taken me longer just to understand how it was done -- and where all the code for it 
was hiding. So it just seemed easier to re-invent it myself. *shrug*