Fr-Sky telemetry GPS screen changes. Large numbers for Altitude, Heading and Speed.
I'm still very confused about the Fr-Sky data protocol. The "after decimal
point" data for example is said to be sent as binary data -- but is it supposed
to be considered as zero-padded or not?

For example, is "9" supposed to be 0.9, 0.09 or 0.009?

In the case of GPS speed for example, I assume two decimal places maximum
(because it sends 99 when satellites are in range but a fix is not yet made). I
guess "25" is supposed to be .25. But is "2" supposed to be .2 or .02?

Alas, it seems my GPS unit is fried. I've received a new Fr-Sky hub, but I still
get no data from the GPS unit. I haven't put the unit itself on a scope to see
if it's doing anything yet. But I guess I'll order a replacmeent soon anyway.