TX_CADDY changes to v1.4.
Essentially, this project is a right, royal FAILURE at this time. However,
the circuitry as presented in this commit _appears_ to function correctly
electrically, I have been unable to get the only available DSMX module
to operate. I am unwillingly to break my working DSM2 module in order
to test that it works -- though I presume it would, if the CRO readings
are to be believed.